New games and minor Fixes (#5)
New games and minor Fixes

+++ PlayGround/PlayGround/Games/Factorial.cs
@@ -10,7 +10,7 @@
             Console.WriteLine($"The factorial is {output}");
         } 
 
-        private static int Fact(int num)
+        private static long Fact(long num)
         {
             if (num ==0 || num == 1)
             {
